% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_where_hit.R
\name{check_where_hit}
\alias{check_where_hit}
\title{Checks where the coordinates of a dart lands on the dartboard in terms of
which scoring wedge it hits.}
\usage{
check_where_hit(x, y, db)
}
\arguments{
\item{x}{is the 'x' coordinate of the shot}

\item{y}{is the 'y' coordinate of the shot}

\item{db}{is a dartbaord object created from the 'dartboard' package.}
}
\value{
the name of the location of where the shot landed (e.g., Treble 20)
}
\description{
Checks where the coordinates of a dart lands on the dartboard in terms of
which scoring wedge it hits.
}
