% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_score_cricket.R
\name{force_score_cricket}
\alias{force_score_cricket}
\title{Forces an update to a player's score without simulating a throw.}
\usage{
force_score_cricket(gm, player_id, bed_val, num_marks, points)
}
\arguments{
\item{gm}{a cricket game object initialized by set_cricket_game}

\item{player_id}{either 1 or 2 to signify which shooter's score will be updated}

\item{bed_val}{vector signifying which bed (20, 19, etc.). Bull = 25 in this case}

\item{num_marks}{vector aligned with bed_val that has the number of marks for the bed}

\item{points}{scalar of the number of point to apply to the shooter}
}
\value{
updated gm, a cricket game object, with the new score
}
\description{
This is useful if you want to stage a specific scenario to simulate.
}
