% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shot.R
\name{plot_shot}
\alias{plot_shot}
\title{Returns a dartboard plot made from dartboard::draw_dartboard() showing targets.}
\usage{
plot_shot(xpt, ypt, db_plot, color)
}
\arguments{
\item{xpt}{vector of "x" coordinates}

\item{ypt}{vector of "y" coordinates}

\item{db_plot}{dartboard::draw_dartboard() object}

\item{color}{is the shot's color}
}
\value{
dartboard::dart_dartboard() object showing shots
}
\description{
Returns a dartboard plot made from dartboard::draw_dartboard() showing targets.
}
