% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_target_basic_cricket.R
\name{select_target_basic_cricket}
\alias{select_target_basic_cricket}
\title{Reasons which shot to select given basic cricket strategy.}
\usage{
select_target_basic_cricket(gm, shooter_player_id, opponent_player_id)
}
\arguments{
\item{gm}{object created from the set_cricket_game}

\item{shooter_player_id}{either 1 or 2 given the shooter}

\item{opponent_player_id}{either 1 or 2, but should not be the shooter's id}
}
\value{
name of the target
}
\description{
Reasons which shot the shooter will select based on a basic cricket strategy.
The shooter
   (1) checks if behind in points
   (2) if not behind, then shoots the highest bed or bull
   (3) if behind, then shooter the highest open bed of the opponent
}
