% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_target_closing_cricket.R
\name{select_target_closing_cricket}
\alias{select_target_closing_cricket}
\title{Reasons which shot to select given a closing strategy.}
\usage{
select_target_closing_cricket(gm, shooter_player_id, opponent_player_id)
}
\arguments{
\item{gm}{object created from the set_cricket_game}

\item{shooter_player_id}{either 1 or 2 given the shooter}

\item{opponent_player_id}{either 1 or 2, but should not be the shooter's id}
}
\value{
name of the target
}
\description{
Reasons which shot the shooter will select based on a closing strategy (aka, boxing).
The shooter
   (1) checks if behind in points (this is modified so that the shooter can be one mark's worth of points behind)
   (2) if not behind, then shoots the highest bed or bull that is open for the opponent
   (3) if not behind and the opponent has closed all marks, then shoot the highest bed or bull
   (4) if behind, then shooter the highest open bed of the opponent
}
