% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cricket_game.R
\name{set_cricket_game}
\alias{set_cricket_game}
\title{Create a cricket game object}
\usage{
set_cricket_game(p1, p2)
}
\arguments{
\item{p1}{player object}

\item{p2}{another player object}
}
\value{
cricket game object
}
\description{
Create a cricket game object
}
