% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_marks.R
\name{set_marks}
\alias{set_marks}
\title{Set the marks and differentials for each player}
\usage{
set_marks(gm)
}
\arguments{
\item{gm}{cricket game object}
}
\value{
updated cricket game object
}
\description{
Set the marks and differentials for each player
}
