% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_player_id.R
\name{set_player_id}
\alias{set_player_id}
\title{Set player id given 1 or 2}
\usage{
set_player_id(id)
}
\arguments{
\item{id}{number for player id, only 1 or 2}
}
\value{
player_id
}
\description{
Set player id given 1 or 2
}
