% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_symbol.R
\name{set_symbol}
\alias{set_symbol}
\title{Turn mark counts into a symbol}
\usage{
set_symbol(number_of_marks)
}
\arguments{
\item{number_of_marks}{vector of the number of marks to turn into a symbol}
}
\value{
vector of symbols
}
\description{
Set symbols for each mark value.
}
\details{
"" for 0 marks
"/" for 1 mark
"X" for 2 marks
"0" for 3 marks
}
