% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_target.R
\name{set_target}
\alias{set_target}
\title{From the target name set the coordinates to shoot}
\usage{
set_target(target, target_dt)
}
\arguments{
\item{target}{name of the target}

\item{target_dt}{data.table of targets created from translate_targets}
}
\value{
cartesian coordiates of the target
}
\description{
From the target name set the coordinates to shoot
}
