% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_game.R
\name{simulate_game}
\alias{simulate_game}
\title{Simulates a cricket game}
\usage{
simulate_game(db, gm, plot_game, verbose = FALSE)
}
\arguments{
\item{db}{dartboard object from dartboard::create_dartboard}

\item{gm}{game object from set_cricket_game}

\item{plot_game}{dartboard plot from dartboard::draw_dartboard}

\item{verbose}{logic to write out data via print statement}
}
\value{
results contain final state of the game and all shots
}
\description{
Simulates a cricket game
}
