% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throw_dart.R
\name{throw_dart}
\alias{throw_dart}
\title{Given a target, finds the actual shot.}
\usage{
throw_dart(target_x, target_y, sd_factor, random_seed, r)
}
\arguments{
\item{target_x}{x coordinate of the target}

\item{target_y}{y coordinate of the target}

\item{sd_factor}{standard deviation}

\item{random_seed}{random seed}

\item{r}{radius of dartboard (to maximum of outer ring)}
}
\value{
coordinates of shot hit
}
\description{
Given a target, finds the actual shot.
}
