% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_targets.R
\name{translate_targets}
\alias{translate_targets}
\title{Translates dartboard into data.table of targets, coordinates, and values.}
\usage{
translate_targets(db)
}
\arguments{
\item{db}{dartboard object}
}
\value{
data.table of target coordinates
}
\description{
Translates dartboard into data.table of targets, coordinates, and values.
}
