% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cricket_game.R
\name{update_game}
\alias{update_game}
\title{Updates cricket object}
\usage{
update_game(gm, shooter_player_id, opponent_player_id, shot_hit,
  targets_dt)
}
\arguments{
\item{gm}{cricket game object}

\item{shooter_player_id}{number either 1 or 2}

\item{opponent_player_id}{number either 1 or 2, but not shooter id}

\item{shot_hit}{where the shot landed}

\item{targets_dt}{data.table of dartboard coordiantes}
}
\value{
udpated game object
}
\description{
Updates cricket object
}
