% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Find the cartesian coordinates of a shot give the name of the target.}
\usage{
get_coordinates(bed_name, bed_mlt)
}
\arguments{
\item{bed_name}{name of the bed (the number of the bed or "Bull")}

\item{bed_mlt}{either "Single", "Double", or "Treble"}
}
\value{
target coordinates given the name of the target
}
\description{
Find the cartesian coordinates of a shot give the name of the target.
}
